requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
langchain==0.1.9
langchain-groq==0.0.1
python-docx==1.1.0
PyPDF2==3.0.1
python-dotenv==1.0.1
tenacity==8.2.2